Change asset to a throw

+++ build_runner/lib/src/asset/file_based.dart
@@ -96,8 +96,10 @@ class FileBasedAssetWriter implements RunnerAssetWriter {
 
   @override
   Future delete(AssetId id) {
-    assert(id.package == packageGraph.root.name,
-        'Only expected to delete assets in ${packageGraph.root.name}: $id');
+    if (id.package != packageGraph.root.name) {
+      throw new InvalidOutputException(
+          id, 'Should not delete assets outside of ${packageGraph.root.name}');
+    }
     if (onDelete != null) onDelete(id);
 
     var file = _fileFor(id, packageGraph);
